requirements.txt
Flask~=3.1.0
SQLAlchemy~=2.0.36
Werkzeug~=3.1.3
python-dateutil~=2.9.0